AUTH-1315 - Move the AwsSqsClient to shared
- We will require the use the AwsSqsClient in the IPV module as we will want to send requests to SPOT via an SQS queue.
