+++ shared/build.gradle
@@ -21,6 +21,7 @@ dependencies {
             configurations.lettuce,
             configurations.hamcrest,
             configurations.sns,
+            configurations.sqs,
             configurations.ssm,
             "com.googlecode.libphonenumber:libphonenumber:8.12.41",
             "com.google.protobuf:protobuf-java:3.19.3"
