+++ shared/src/main/java/uk/gov/di/authentication/shared/services/AwsSqsClient.java
@@ -1,0 +1,39 @@
+package uk.gov.di.authentication.shared.services;
+
+import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
+import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
+import software.amazon.awssdk.core.exception.SdkClientException;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.sqs.SqsClient;
+import software.amazon.awssdk.services.sqs.SqsClientBuilder;
+import software.amazon.awssdk.services.sqs.model.SendMessageRequest;
+
+import java.net.URI;
+import java.util.Optional;
+
+public class AwsSqsClient {
+
+    private final SqsClient client;
+    private final String queueUrl;
+
+    public AwsSqsClient(String region, String queueUrl, Optional<String> sqsEndpoint) {
+        SqsClientBuilder amazonSqsBuilder = SqsClient.builder().region(Region.of(region));
+
+        if (sqsEndpoint.isPresent()) {
+            amazonSqsBuilder
+                    .endpointOverride(URI.create(sqsEndpoint.get()))
+                    .credentialsProvider(
+                            StaticCredentialsProvider.create(
+                                    AwsBasicCredentials.create("FAKEACCESSKEY", "FAKESECRETKEY")));
+        }
+        this.client = amazonSqsBuilder.build();
+        this.queueUrl = queueUrl;
+    }
+
+    public void send(final String event) throws SdkClientException {
+        SendMessageRequest messageRequest =
+                SendMessageRequest.builder().queueUrl(queueUrl).messageBody(event).build();
+
+        client.sendMessage(messageRequest);
+    }
+}
